requirements.txt
asgiref==3.7.2
black==24.2.0
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==5.0.2
django-dynamic-admin-forms==2.2.1
django-environ==0.11.2
django-import-export==3.3.7
django-unfold==0.20.4
et-xmlfile==1.1.0
gunicorn==21.2.0
h11==0.14.0
MarkupPy==1.14
mypy-extensions==1.0.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
